+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Versions 0.5.10b
+* Adjusted feat tools to better handle metamagic.
+* Support for Oracle Burning Magic.
+
 ## Versions 0.5.10a
 * Added fallthrough case for PrerequisiteDeityAlignment in case someone breaks deity alignment checks.